requirements.txt
# === DATABASE & AUTHENTICATION ===
sqlalchemy
passlib
bcrypt=4.3.0
pydantic
python-jose
# === BACKEND ===
fastapi==0.116.1
uvicorn==0.27.1
tensorflow==2.20.0
joblib==1.3.2
scikit-learn==1.8.0

# === FRONTEND ===
streamlit==1.36.0
requests==2.31.0
pillow==12.1.0

# === COMMON ===
numpy==2.4.1